renamed gimbtononmy to neopixel
