--- uwrt_mars_rover_gimbtonomy/src/uwrt_mars_rover_gimbtonomy_node.cpp
@@ -1,10 +1,0 @@
-#include <cstdlib>
-
-#include "uwrt_mars_rover_gimbtonomy/uwrt_mars_rover_gimbtonomy_can.h"
-#include "uwrt_mars_rover_gimbtonomy/uwrt_mars_rover_gimbtonomy_neopixel.h"
-
-int main(int argc, char** argv){
-    neopixel neopixel_obj(argc, argv);
-    neopixel_obj.run();
-    return EXIT_SUCCESS;
-}
