+++ uwrt_mars_rover_neopixel/src/uwrt_mars_rover_neopixel_node.cpp
@@ -1,0 +1,10 @@
+#include <cstdlib>
+
+#include "uwrt_mars_rover_neopixel/uwrt_mars_rover_neopixel_can.h"
+#include "uwrt_mars_rover_neopixel/uwrt_mars_rover_neopixel_control.h"
+
+int main(int argc, char** argv){
+    neopixel neopixel_obj(argc, argv);
+    neopixel_obj.run();
+    return EXIT_SUCCESS;
+}
